Add files via upload
This "Tutor" web app is made for our database project. We used mysql as our database and jsp for backend. For frontend we used html, css, bootstrap. This is a platform for students who want to find a tution class for their subject. For the phase 1 we didnt consider the students whereabouts and filter the ads. Teachers and Institutes can post ads and student can decide which one is the best for them.